Alternate graph code:
Now can specify an alternate path in the config file for experiment.perl
All you need to do is change the line: decoding-steps
Every : you add will mean that the next steps are in an alternate graph
Ie. 
	decoding-steps = " t0 , g1 , t1 "
will produce a single graph but
	decoding-steps = " t0 , g1 : t1 "
will produce two alternate graphs

The only change it actually makes is to the create configuration file.

This is also a bugfix!
There was no ability to have multiple factors in the target
for the two filering scripts. They just ignored those lines
as if they were not valid table descriptions.




git-svn-id: http://svn.statmt.org/repository/mira@1322 cc96ff50-19ce-11e0-b349-13d7f0bd23df