- Bug gix: lm.log cannot be written.
The bjam script would return with:
failed to write output file '/path/to/src/moses/src/LM/bin/lm.log'!
Seems either directory bin/ is missing or this was not the expected target. Seeing a bin/ at the root,
and a config.log inside, I went with the assumption that lm.log was supposed to go there as well.