Fix bug on MERT Perl script when working directory is not specified.
If you set a working directory from the command line option,
the script works fine.

However, if you don't, the script will crash when calling a function
`create_extractor_script()' which creates shell script for the following
MERT iterations due to the inconsistency between path to default working
directory and Perl interpreter's current directory.

So, We assume that the default working directory used in the script
will be set to "`pwd`/mert-work".