Fix a mistake in a previous commit: tuning on a subset of features.
In the commit 4b6232b757542e7dfd05b6c494d5f97fdb80c448,
I thought I had fixed the bug around the tuning on a subset of
features by checking whether pdim and the length of the
active features which you want to optimize in the tuning.

However, it was wrong. I should set Point::optindices
appropriately according to specified the subset.