FIXED OPTION FOR FUTURE DISTORTION COST [Moore & Quirk 2007]
Summary of changes

1) DistortionScoreProducer::CalculateDistortionScore

The previous implementation had the following bugs:
- wasn't correctly converting size_t to int
- in initial (empty) hypothesis, prefixEndPos was -2 instead of -1
- nb of words between phrases was always one too much

2) DistortionScoreProducer::Evaluate

The new distortion state was assgined the first gap of the old hypothesis (it should be the current's)

3) WordsRange::GetNumWordsBetween

It returned one word too much. For instance the nb of words between [0..1] and [2..3] was 1, now it's 0.

4) Parameter.cpp, StaticData.cpp, StaticData.h

Added binary option to activate future distortion cost (fdc) and corresponding StaticData's variable