truecase corpus before training recaser
gives better results in (small) test, and the code already had a placeholder for it.
(without truecasing, the recaser is more likely to uppercase words like "the" if they are often sentence-initial in the training corpus)

If people don't want the default behavior changed, I can disable truecasing by default and add a command line parameter to enable it.