Support for compact phrase table in EMS
It should be sufficient to add a line like
ttable-binarizer = "/home/bhaddow/moses/dist/bin/processPhraseTableMin"
to your EMS config, and everything else will be taken care of. You can
add other arguments to the processPhraseTableMin, for example for
threading, by putting them in the quotes.

Note that this is not fully tested, since there are currently some
issues with the compact phrase table introduced by the sparse feature
merge.