Fixes compilation error using libboost_1_50
Problem spotted in https://github.com/moses-smt/mosesdecoder/issues/32 
fixed.

According to the Boost doc, nbestPath.parent_path().filename() 
returns a path object, to get the correspondent std::string 
representation, one must call one of the methods listed in:
http://www.boost.org/doc/libs/1_53_0/libs/filesystem/doc/reference.html#path-native-format-observers

native() is supposed to return the path in the specific OS
path format (using backslashes for Windows). Anyway, since we 
are considering only the filename here, the result is the same.