Prevented segfault on print non-existant alignment
In processPhraseTable, there is an option to print alignments.
When this is selected, processPhraseTable will print the 4th
column in the phrase table, but the check for the phrase table
width only currently goes to the 3rd column. This patch makes
sure that the phrase table must have 4 columns when alignments
are being printed, and 3 otherwise.