Compare nullness in Hypothesis::RecombineCompare.
The comparison code subtracted two pointers in the case where at least one
was null, to get a signed int comparison result.  But that subtraction is
undefined, and the cast to int (dropping the most-significant bits!) made the
outcome even less uncertain.

In this patch I compare the nullness of the pointers instead, which should
always return a well-defined -1, 0, or 1.